changed prepareRenderPath()
setting the path must be central in main config. To be free where you want to write the files, the function must not prefix the path with some basepath at all. For backward compatibility the old params remain and new params are introduced:
- publicRuntimeAlias
- protectedRuntimeAlias
These params hold (not the filesystem paths anymore but) the Yii alias of the paths.